requirements.txt
einops==0.8.0
torch==2.3.1+cu121
torchvision==0.18.1+cu121
lpips==0.1.4
matplotlib==3.7.5
numpy==1.24.1
pandas==2.2.2
Pillow==10.4.0
Requests==2.32.3
scikit_learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
tensorboardX==2.6.2.2
timm==1.0.9
tqdm==4.66.4
mamba_ssm==2.1.0
